feat: `#guard_expr` command and adjust handling of expected types for `guard_expr` tactics (#133)
This adds a `#guard_expr` command that behaves just like the `guard_expr` tactic, but it's a command rather than a tactic.

Furthermore, this adjusts the handling of expected types so that when synthetic metavariables are synthesized then instances are more likely to be correct. This affects numeric literals for example, which otherwise might become `Nat`s due to the default instance.

Lastly, adds a `#guard` command that uses the evaluator to check if the given expression evaluates to true.

This PR is moving commands added for testing in [mathlib4#3427](https://github.com/leanprover-community/mathlib4/pull/3427) to std4.