feat: DiscrTree: Avoid unsafe/partial definitions
in #285 there was a comment that it is desirable to avoid `unsafe` and
`partial` definitions here. This PR explores making these definitions
not `unsafe` by proving termiation.

This requires `Array.attach`, analogous to `List.attach`. Futhermore,
because the recursion also goes throuh another layer of `Prod`, a custom
lemma and recursion tactic seems to be require. I conjecture that a more
compositional default `recursion_tactic` setup might make that not
necessary.

The function `mergePreservingDuplicates` is particularly tricky, as it
recurses through `Array.mergeSortedMergingDuplicates`, which I have to
generalize to make this work. This examples show a benefit of extrinsic
termination proofs a la Isabelle (where one would have to prove a
separate `congr` lemma for `Array.mergeSortedMergingDuplicates`, but
would not have to change its definition), althought I am not yet sure if
such an approach to well-founded recursion is possible in Lean.