feat: avoid termination_by', introduce WellFounded.wrap
I’d like to remove support for the `termination_by'` annotation. Until
https://github.com/leanprover/std4/pull/371 it wasn't used anywhere in
lean, std, mathlib, so this PR removes the single use of it.

It does so using the pattern that can be used to replace uses of
`termination_by'`, should there be more: Using the helper type
`WellFounded.Wrap` one can indicate an explicit `WellFounded` relation
to use.

So this PR uses that pattern to avoid the use of `termination_by'` here,
and at the same time provides the necessary definitions for others, so
when Lean drops support for `termination_by'`
(https://github.com/leanprover/lean4/pull/3033), we can tell users how
migrate.