fix: don't panic when linting in empty environments (#494)
The "unreachable tactic" linter would previously assume that various
parser categories were in the environment, and panic if they
weren't. This caused problems when elaborating declarations in a
fresh, empty environment.

This also adds a regression test that makes sure linters work in a fresh
environment and configures testing to fail if Lean panics.