feat: allow configuring additional linters to run
Extend getChecks to allow an additional list of linters, which is always run.
This can be used downstream to e.g. enable the docBlameThm or explicitVarsOfIff
linters without modifying batteries.

As part of this, we
- update and fix the function doc-string (which was quite outdated)
- remove the useOnly option: the equivalent effect can be achieved by
filtering the list of linters afterwards; *adding* linters cannot,
hence must be performed in this function
- rename a variable `dflt` to the more readable `default`