Mark `Nat.sqrt.iter` as semireducible
This definition has become irreducible in v4.19.0-rc2, which blocks reduction of `Nat.sqrt`, as well as `IsSquare` and `Nat.minFac` in mathlib, both functions which are used for decidability there. Furthermore, marking `Nat.sqrt.iter` as `semireducible` locally does not make those reduce appropriately. Further still, `decide +kernel` - which is meant to ignore irreducible tags - fails to see through `Nat.sqrt.iter`, so the only solution is this one.