feat: ascii case insensitive Char/String compare
1. Adds ascii case insensitive comparison to `Char` and `String`.

2. Adds a type of ascii case insensitive strings. This is more likely to be
controversial and I'm not wed to it, but this concept is pervasive in
basic networking code that seems like the kind of thing that Batteries
seeks to enable. Case in point, every nascent attempt at implementing
HTTP and URI listed on reservoir reimplements their own version of this
type.

Concerning the admittedly verbose type name:

+ Variants of the more direct `CaseInsensitiveString` were rejected for
clarity, as this type does not do more complex locale-based UTF-8
comparison.

+ `AsciiCaseInsensitiveString` was passed over because it might be
interpreted as "ascii string, case insensitive", while the string itself
is still UTF8.