feat: make `unusedHavesSuffices` linter only apply to propositions
The `unusedHavesSuffices` linter error talks about proofs, so this PR adds in a check that the `have` is indeed introducing a proposition. This will be necessary in the next release because Lean will convert `let`s to `have`s automatically to cache which ones are nondependent, and, for example, `do` notation can elaborate into expressions that have unused `let`s, which we don't want to be flagged by this linter.

After the next Lean release, this linter should be updated look for nondependent `Expr.letE` expressions rather than `letFun`s.