feat: upstream the `library_note` command from Mathlib
This PR moves Mathlib's `library_note2` command to Batteries, replacing
`library_note`. The main difference, apart from a slightly changed
syntax, is `library_note` now adds a declaration to the `LibraryNote`
namespace.

As a consequence, it is no longer possible to have two library notes
with the same name. (Was this feature ever used?)